fix(layout): black video on screencasting (chrome)
The issue is that for some reason, even when the stream is available and the isBigVideoVisible is true, the reference to bigVideo ViewChild (that should be the #bigVideo video element) remains undefined.
This happens using chrome and starting a screencasting video session when the customer hasn't yet interacted with the page.
Using hidden attribute instead of destory-create the bigVideo parent div with ngIf allows the viewChild to be always available.